spark-submit insists all conf is prefixed 'spark.'
The _simplest thing that could possibly work™_ is to require the prefix on all
our own configuration keys. This is the approach taken by Apache Iceberg. The
purist in me was thinking to first reference the keys by prefixing
"com.xebia.[...]" and only if that comes up empty, trying again with the
"spark.com.xebia.[...]" prefix. This was rejected for not being the simplest
thing that could possibly work.

A future version could take this chained approach, and also look for JVM System
Properties and environment variables as configuration sources. For now, this
will do just fine.